ClickAid Java Applet
A game in Java Applet. A user define nxn grid is generated, each button
represent a country and the number inside them represent their mistery
index.  The buttons are counties and the numbers from 0-9 is a
representation of misery index of the country. Clicking a country
provides a aid package decresing the misery index of the country and its
neighbours by 1 but providing an aid package to a country with 0 index
will backfire and its misery index will be increased to number of non
zero countries of its neighbours.